addNumberElement
Purpose for this: When you try to put a number by string into object, and take it out by converting into string, it will include the quotes from the object. Then when it converts into number, it returns undefined. It leads to more complicated scripting. Therefore I thought about this idea.
(This is my first time for making this stuff so I am still not familiar how does it work)